backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
openai==1.3.7
pandas==2.1.3
numpy==1.25.2
redis==5.0.1
celery==5.3.4
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
openpyxl==3.1.2
xlrd==2.0.1
psycopg2-binary==2.9.9
